requirements.txt
Flask==2.3.2
python-telegram-bot==20.3
httpx==0.25.2